chore: mysql 8.0.33 (#7284)
* chore: mysql 8.0.33

* chore: update

Former-commit-id: 759540ff39f2394f5a359b3bfd82d0624bffcc57
Former-commit-id: 84bb751ae638d90e418dc65686648d3c014d60b0

+++ backend/resources/mysql/fetch_mysql.sh
@@ -12,10 +12,12 @@ atomic_download() {
     if [[ ! -e $1 ]]; then
         echo "Downloading $1..."
         curl -o $1.tmp https://cdn.mysql.com/archives/mysql-8.0/$1
-        if $(ckmd5 $1.tmp $2); then
-            mv $1.tmp $1
-        else
-            rm $1.tmp
+        if [ -n "$2" ]; then 
+            if $(ckmd5 $1.tmp $2); then
+                mv $1.tmp $1
+            else
+                rm $1.tmp
+            fi
         fi
     fi
 }
